refactor: create a reusable code editor component
