+++ packages/twenty-front/src/modules/settings/serverless-functions/components/tabs/SettingsServerlessFunctionTestTab.tsx
@@ -2,6 +2,7 @@ import { Section } from '@/ui/layout/section/components/Section';
 import { H2Title, IconPlayerPlay } from 'twenty-ui';
 
 import { LightCopyIconButton } from '@/object-record/record-field/components/LightCopyIconButton';
+import { SettingsServerlessFunctionCodeEditorContainer } from '@/settings/serverless-functions/components/SettingsServerlessFunctionCodeEditorContainer';
 import { SettingsServerlessFunctionsOutputMetadataInfo } from '@/settings/serverless-functions/components/SettingsServerlessFunctionsOutputMetadataInfo';
 import { settingsServerlessFunctionCodeEditorOutputParamsState } from '@/settings/serverless-functions/states/settingsServerlessFunctionCodeEditorOutputParamsState';
 import { settingsServerlessFunctionInputState } from '@/settings/serverless-functions/states/settingsServerlessFunctionInputState';
@@ -78,37 +79,31 @@ export const SettingsServerlessFunctionTestTab = ({
               />,
             ]}
           />
-          <CodeEditor
-            files={[
-              {
-                content: settingsServerlessFunctionInput,
-                language: 'json',
-                path: 'input.json',
-              },
-            ]}
-            currentFilePath={'input.json'}
-            height={200}
-            onChange={setSettingsServerlessFunctionInput}
-          />
+          <SettingsServerlessFunctionCodeEditorContainer>
+            <CodeEditor
+              value={settingsServerlessFunctionInput}
+              language="json"
+              height={200}
+              onChange={setSettingsServerlessFunctionInput}
+              options={{ readOnly: true, domReadOnly: true }}
+            />
+          </SettingsServerlessFunctionCodeEditorContainer>
         </div>
         <div>
           <CoreEditorHeader
             leftNodes={[<SettingsServerlessFunctionsOutputMetadataInfo />]}
             rightNodes={[<LightCopyIconButton copyText={result} />]}
           />
-          <CodeEditor
-            files={[
-              {
-                content: result,
-                language:
-                  settingsServerlessFunctionCodeEditorOutputParams.language,
-                path: 'result.any',
-              },
-            ]}
-            currentFilePath={'result.any'}
-            height={settingsServerlessFunctionCodeEditorOutputParams.height}
-            options={{ readOnly: true, domReadOnly: true }}
-          />
+          <SettingsServerlessFunctionCodeEditorContainer>
+            <CodeEditor
+              value={result}
+              language={
+                settingsServerlessFunctionCodeEditorOutputParams.language
+              }
+              height={settingsServerlessFunctionCodeEditorOutputParams.height}
+              options={{ readOnly: true, domReadOnly: true }}
+            />
+          </SettingsServerlessFunctionCodeEditorContainer>
         </div>
       </StyledInputsContainer>
     </Section>
